Use timeouts

+++ test/integration/activate_int_test.go
@@ -1,6 +1,7 @@
 package integration
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"os/exec"
@@ -11,6 +12,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ActiveState/cli/internal/rtutils"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/ActiveState/cli/internal/constants"
@@ -68,13 +70,16 @@ func (suite *ActivateIntegrationTestSuite) addForegroundSvc(ts *e2e.Session) fun
 	suite.Require().NoError(err)
 
 	// Wait for the svc to be ready
-	code := -1
-	for code != 0 {
-		code, _, _ = exeutils.Execute(ts.SvcExe, []string{"status"}, func(cmd *exec.Cmd) error {
-			cmd.Env = ts.Env
-			return nil
-		})
-	}
+	rtutils.Timeout(func() error {
+		code := -1
+		for code != 0 {
+			code, _, _ = exeutils.Execute(ts.SvcExe, []string{"status"}, func(cmd *exec.Cmd) error {
+				cmd.Env = ts.Env
+				return nil
+			})
+		}
+		return nil
+	}, 10*time.Second)
 
 	// Stop function
 	return func() {
@@ -83,18 +88,13 @@ func (suite *ActivateIntegrationTestSuite) addForegroundSvc(ts *e2e.Session) fun
 			suite.Require().NoError(err, "svc stop failed: %s\n%s", stdout, stderr)
 		}()
 
-		errCh := make(chan error)
-		go func() {
-			errCh <- cmd.Wait()
-		}()
-
 		verifyExit := true
 
-		var err error
-		select {
-		case err = <-errCh:
-			break
-		case <-time.After(10 * time.Second):
+		err2 := rtutils.Timeout(func() error { return cmd.Wait() }, 10*time.Second)
+		if err2 != nil {
+			if !errors.Is(err2, rtutils.ErrTimeout) {
+				suite.Require().NoError(err2)
+			}
 			fmt.Printf("svc did not stop in time, Stdout:\n%s\n\nStderr:\n%s", stdout.String(), stderr.String())
 			cmd.Process.Kill()
 
@@ -105,12 +105,11 @@ func (suite *ActivateIntegrationTestSuite) addForegroundSvc(ts *e2e.Session) fun
 			// We'll address this properly with the refactor: DX-1312
 			// All that said; we should still be able to verify the output, which is the real meat of this function anyway.
 			verifyExit = false
-			break
 		}
 
 		errMsg := fmt.Sprintf("svc foreground did not complete as expected. Stdout:\n%s\n\nStderr:\n%s", stdout.String(), stderr.String())
 		if verifyExit {
-			suite.Require().NoError(err, errMsg)
+			suite.Require().NoError(err2, errMsg)
 			if cmd.ProcessState.ExitCode() != 0 {
 				suite.FailNow(errMsg)
 			}
